+++ ouroboros-network/src/Ouroboros/Network/Diffusion/Types.hs
@@ -128,6 +128,17 @@ data Tracers ntnAddr ntnVersion ntnVersionData
       dtMuxTracer
         :: Tracer m (Mx.WithBearer (ConnectionId ntnAddr) Mx.Trace)
 
+      -- | Mux's channel tracer
+      --
+    , dtChannelTracer
+        :: Tracer m (Mx.WithBearer (ConnectionId ntnAddr) Mx.ChannelTrace)
+
+      -- | Bearer tracer
+      --
+    , dtBearerTracer
+        :: Tracer m (Mx.WithBearer (ConnectionId ntnAddr) Mx.BearerTrace)
+
+
       -- | Handshake protocol tracer
     , dtHandshakeTracer
         :: Tracer m (NodeToNode.HandshakeTr ntnAddr ntnVersion)
@@ -137,8 +148,19 @@ data Tracers ntnAddr ntnVersion ntnVersionData
       --
 
       -- | Mux tracer for local clients
+      --
     , dtLocalMuxTracer
         :: Tracer m (Mx.WithBearer (ConnectionId ntcAddr) Mx.Trace)
+
+      -- | Mux's channel tracer for local clients
+      --
+    , dtLocalChannelTracer
+        :: Tracer m (Mx.WithBearer (ConnectionId ntcAddr) Mx.ChannelTrace)
+
+      -- | Bearer tracer for local clients
+      --
+    , dtLocalBearerTracer
+        :: Tracer m (Mx.WithBearer (ConnectionId ntcAddr) Mx.BearerTrace)
 
       -- | Handshake protocol tracer for local clients
     , dtLocalHandshakeTracer
@@ -227,8 +249,12 @@ nullTracers :: Applicative m
                        extraFlags extraPeers extraCounters m
 nullTracers = Tracers {
     dtMuxTracer                                  = nullTracer
+  , dtChannelTracer                              = nullTracer
+  , dtBearerTracer                               = nullTracer
   , dtHandshakeTracer                            = nullTracer
   , dtLocalMuxTracer                             = nullTracer
+  , dtLocalChannelTracer                         = nullTracer
+  , dtLocalBearerTracer                          = nullTracer
   , dtLocalHandshakeTracer                       = nullTracer
   , dtDiffusionTracer                            = nullTracer
   , dtTraceLocalRootPeersTracer                  = nullTracer