+++ ouroboros-network/testlib/Test/Ouroboros/Network/Diffusion/Node.hs
@@ -301,6 +301,7 @@ run blockGeneratorArgs limits ni na
               , Diffusion.daNtnHandshakeArguments =
                   HandshakeArguments
                     { haHandshakeTracer  = nullTracer
+                    , haBearerTracer     = nullTracer
                     , haHandshakeCodec   = unversionedHandshakeCodec
                     , haVersionDataCodec = ntnUnversionedDataCodec
                     , haAcceptVersion    = iAcceptVersion ni
@@ -310,6 +311,7 @@ run blockGeneratorArgs limits ni na
               , Diffusion.daNtcHandshakeArguments =
                   HandshakeArguments
                     { haHandshakeTracer  = nullTracer
+                    , haBearerTracer     = nullTracer
                     , haHandshakeCodec   = unversionedHandshakeCodec
                     , haVersionDataCodec = unversionedProtocolDataCodec
                     , haAcceptVersion    = \_ v -> Accept v